+++ lazyapp/telemetry_enabled_lazydev.go
@@ -1,0 +1,9 @@
+//go:build lazydev
+
+package lazyapp
+
+import "golazy.dev/lazytelemetry"
+
+func telemetryMiddlewareEnabled(config lazytelemetry.Config) bool {
+	return !config.SDKDisabled
+}