+++ lazytelemetry/middleware_logger_lazydev.go
@@ -1,0 +1,15 @@
+//go:build lazydev
+
+package lazytelemetry
+
+import (
+	"io"
+	"log/slog"
+)
+
+func defaultMiddlewareLogger(config Config) *slog.Logger {
+	if config.Enabled() {
+		return NewLogger(config, nil)
+	}
+	return NewLogger(config, io.Discard)
+}