+++ lazytelemetry/middleware_logger_not_lazydev.go
@@ -1,0 +1,9 @@
+//go:build !lazydev
+
+package lazytelemetry
+
+import "log/slog"
+
+func defaultMiddlewareLogger(config Config) *slog.Logger {
+	return NewLogger(config, nil)
+}